+++ src/extension.ts
@@ -1,6 +1,7 @@
 import * as vscode from 'vscode';
 import { ZenViewProvider } from './ZenViewProvider';
 import { zenViewGlobals } from './ZenViewGlobals';
+import { ConfigHandler, FilePathType } from './ConfigHandler';
 
 const zenViewProvider = new ZenViewProvider();
 
@@ -11,11 +12,23 @@ export function activate(context: vscode.ExtensionContext) {
     return;
   }
   vscode.window.registerTreeDataProvider('zenView', zenViewProvider);
-  vscode.commands.registerCommand('zenView.open', (fileUri) => vscode.commands.executeCommand('vscode.open', fileUri));
-	vscode.workspace.onDidChangeConfiguration(handleConfigChangeEvent);
+  vscode.commands.registerCommand('zenView.open', (fileUri) => {
+    vscode.commands.executeCommand('vscode.open', fileUri);
+  });
+  vscode.commands.registerCommand('zenView.addRelativePath', (fileUri: vscode.Uri) => {
+    ConfigHandler.addZenPath(rootPath!, fileUri, FilePathType.relative);
+    onConfigChange();
+  });
+  vscode.commands.registerCommand('zenView.addAbsolutePath', (fileUri: vscode.Uri) => {
+    ConfigHandler.addZenPath(rootPath!, fileUri, FilePathType.absolute);
+    onConfigChange();
+  });
+  vscode.workspace.onDidChangeConfiguration(() => {
+    onConfigChange();
+  });
 }
 
-function handleConfigChangeEvent() {
+function onConfigChange() {
   zenViewGlobals.onConfigChange();
   zenViewProvider.refresh();
 }