+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 All notable changes to the "zenview" extension will be documented in this file.
 
-## [1.0.0] - 2022-01-23
-### Added## 
+## [1.1.0] - 2022-01-24
+### Added
+- Context menu in file explorer to add paths as absolute or relative path
 
+### Fixed
+- Fixed directories not using the correct icon from the icon theme in the zen view
+
+## [1.0.0] - 2022-01-23
+### Added
 - Initial release